# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCRM API",
    "description": "API REST du CRM interne DCB Technologies. Format de réponse unifié : succès `{ success: true, data, meta? }`, erreur `{ success: false, error: { code, message } }`. Authentification par JWT Bearer (15 min, refresh 7 j via cookie httpOnly) ou par clé API longue durée (header X-API-Key). L'extension `x-permission` sur une opération indique la clé de permission RBAC requise. Tous les montants monétaires sont exprimés en euros hors taxes (HT) ; le TTC n'est jamais stocké et se dérive via Product.vatRate (défaut 20 %).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Développement"
    },
    {
      "url": "https://{domaine}/api",
      "description": "Production",
      "variables": {
        "domaine": {
          "default": "crm.example.fr"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtenu via POST /auth/login (durée 15 min)"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clé API longue durée générée via POST /apikeys (préfixe dcrm_). Permissions effectives = intersection des droits de la clé (scopes) et des droits actuels de son propriétaire — jamais de bypass '*', même pour un ADMIN ; les routes gardées par rôle sont inaccessibles par clé API."
      }
    },
...
  },
  "paths": {
...
    "/settings": {
      "get": {
        "tags": [
          "Paramètres"
        ],
        "summary": "Liste tous les paramètres (fusionnés avec les valeurs par défaut)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "settings:read",
        "responses": {
          "200": {
            "description": "data: [{ key, value, label }]"
          }
        }
      }
    },
    "/settings/{key}": {
      "get": {
        "tags": [
          "Paramètres"
        ],
        "summary": "Récupère un paramètre",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "settings:read",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ key, value, label }"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "put": {
        "tags": [
          "Paramètres"
        ],
        "summary": "Met à jour un paramètre (clés d'authentification Google réservées au rôle ADMIN)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "settings:write",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "value"
                ],
                "properties": {
                  "value": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setting upserté"
          },
          "400": {
            "description": "INVALID_DOMAIN / INVALID_ROLE / VALIDATION_ERROR"
          },
          "403": {
            "description": "FORBIDDEN — clé sensible (googleAllowedDomain/googleAutoCreateRole), réservée à un ADMIN"
          },
          "404": {
            "description": "NOT_FOUND — clé hors DEFAULTS"
          }
        }
      }
    },
    "/settings/actions/run-contract-update": {
      "post": {
        "tags": [
          "Paramètres"
        ],
        "summary": "Déclenche manuellement le job de mise à jour des statuts de contrats",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "settings:write",
        "responses": {
          "200": {
            "description": "data: résultat brut de runContractStatusUpdate() (voir scheduler.ts)"
          }
        }
      }
    },
...
  }
}