# server/src/routes/settings.ts
import { Router, Response } from 'express'
import { z } from 'zod'
import prisma from '../prisma/client'
import { authenticate, AuthRequest, requirePermission } from '../middleware/auth'
import { handleRouteError } from '../middleware/errorHandler'
import { restartScheduler } from '../scheduler'

const router = Router()
router.use(authenticate)

// Default settings values
const DEFAULTS: Record<string, { value: string; label: string }> = {
  contractExpiringSoonDays: { value: '60',      label: 'Alerte contrats expirant dans (jours)' },
  licenseExpiringSoonDays:  { value: '30',      label: 'Alerte licences expirant dans (jours)' },
  warrantyExpiringSoonDays: { value: '60',      label: 'Alerte garanties équipement expirant dans (jours)' },
  schedulerEnabled:         { value: 'true',    label: 'Mise à jour automatique des statuts activée' },
  schedulerTime:            { value: '02:00',   label: 'Heure d\'exécution du job (HH:MM, heure serveur)' },
  companyName:              { value: 'DCB Technologies',  label: 'Nom de l\'entreprise' },
  companyLogoUrl:           { value: '',        label: 'URL du logo' },
  companyAddress:           { value: '',        label: 'Adresse' },
  companyContactEmail:      { value: '',        label: 'Email de contact' },
  companyPhone:             { value: '',        label: 'Téléphone' },
  companySiret:             { value: '',        label: 'SIRET' },
  companyVatNumber:         { value: '',        label: 'N° TVA' },
  callRecordingRetentionDays: { value: '180',  label: 'Rétention des enregistrements d\'appels (jours)' },
  slaHoursCritical:         { value: '4',       label: 'SLA tickets — priorité critique (heures)' },
  slaHoursHigh:             { value: '8',       label: 'SLA tickets — priorité élevée (heures)' },
  slaHoursNormal:           { value: '24',      label: 'SLA tickets — priorité normale (heures)' },
  slaHoursLow:              { value: '72',      label: 'SLA tickets — priorité faible (heures)' },
  googleAllowedDomain:        { value: 'dcb-technologies.fr', label: 'Domaine email autorisé pour l\'auto-création via Google OAuth' },
  googleAutoCreateRole:       { value: 'COMMERCIAL', label: 'Rôle attribué lors de l\'auto-création d\'un compte via Google OAuth' },
}

// GET /api/settings — all settings (settings:read)
router.get('/', requirePermission('settings:read'), async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const rows = await prisma.setting.findMany()
    // Merge DB values with defaults so all keys are always present
    const result = Object.entries(DEFAULTS).map(([key, def]) => {
      const row = rows.find(r => r.key === key)
      return { key, value: row?.value ?? def.value, label: def.label }
    })
    res.json({ success: true, data: result })
  } catch (err) { handleRouteError(err, res) }
})

// GET /api/settings/:key — single setting (internal use, no role restriction)
router.get('/:key', requirePermission('settings:read'), async (req: AuthRequest, res: Response): Promise<void> => {
  try {
    const { key } = req.params
    const row = await prisma.setting.findUnique({ where: { key } })
    const def = DEFAULTS[key]
    const value = row?.value ?? def?.value ?? null
    if (value === null) { res.status(404).json({ success: false, error: { code: 'NOT_FOUND', message: 'Paramètre introuvable' } }); return }
    res.json({ success: true, data: { key, value, label: def?.label } })
  } catch (err) { handleRouteError(err, res) }
})

/**
 * Réglages qui gouvernent l'authentification : ils décident qui peut obtenir un compte.
 * `googleAllowedDomain` est l'unique garde-fou de l'auto-création via Google OAuth
 * (routes/auth.ts) — le laisser derrière `settings:write`, que MANAGER possède, revenait
 * à transformer une permission de paramétrage en provisioning de comptes non authentifié
 * (pointer le domaine sur un fournisseur public ouvre l'inscription à tout Internet).
 */
const AUTH_SENSITIVE_KEYS = new Set(['googleAllowedDomain', 'googleAutoCreateRole'])

/** Nom de domaine simple : labels alphanumériques séparés par des points, TLD d'au moins 2 lettres. */
const DOMAIN_RE = /^(?=.{1,253}$)[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?(\.[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?)*\.[a-z]{2,}$/

// PUT /api/settings/:key — update setting (les clés d'authentification sont réservées aux ADMIN)
router.put('/:key', requirePermission('settings:write'), async (req: AuthRequest, res: Response): Promise<void> => {
  try {
    const { key } = req.params
    const { value } = z.object({ value: z.string() }).parse(req.body)
    if (!DEFAULTS[key]) { res.status(404).json({ success: false, error: { code: 'NOT_FOUND', message: 'Paramètre inconnu' } }); return }

    if (AUTH_SENSITIVE_KEYS.has(key) && !req.permissions?.includes('*')) {
      res.status(403).json({ success: false, error: { code: 'FORBIDDEN', message: 'Seul un administrateur peut modifier les paramètres de connexion Google' } })
      return
    }

    let finalValue = value

    // Domaine autorisé pour l'auto-création Google : format vérifié et normalisé en minuscules.
    // Sans validation, une valeur vide ou un caractère parasite ouvrait ou cassait silencieusement
    // le contrôle de domaine côté callback OAuth.
    if (key === 'googleAllowedDomain') {
      const domain = value.trim().toLowerCase().replace(/^@/, '')
      if (!DOMAIN_RE.test(domain)) {
        res.status(400).json({ success: false, error: { code: 'INVALID_DOMAIN', message: 'Domaine invalide (attendu : exemple.fr)' } })
        return
      }
      finalValue = domain
    }

    // Le rôle d'auto-création Google ne doit jamais être ADMIN (ADMIN = bypass total des permissions).
    // On vérifie aussi que le rôle existe, et on normalise en majuscules (les noms de rôles le sont).
    if (key === 'googleAutoCreateRole') {
      const roleName = value.trim().toUpperCase()
      if (roleName === 'ADMIN') {
        res.status(400).json({ success: false, error: { code: 'INVALID_ROLE', message: 'Le rôle d\'auto-création ne peut pas être ADMIN' } })
        return
      }
      const roleExists = await prisma.role.findUnique({ where: { name: roleName } })
      if (!roleExists) {
        res.status(400).json({ success: false, error: { code: 'INVALID_ROLE', message: 'Rôle inconnu' } })
        return
      }
      finalValue = roleName
    }

    const setting = await prisma.setting.upsert({
      where: { key },
      update: { value: finalValue, label: DEFAULTS[key].label },
      create: { key, value: finalValue, label: DEFAULTS[key].label },
    })
    // Relancer le scheduler si un paramètre lié est modifié
    if (['schedulerEnabled', 'schedulerTime'].includes(key)) {
      restartScheduler().catch(console.error)
    }
    res.json({ success: true, data: setting })
  } catch (err) { handleRouteError(err, res) }
})

// POST /api/settings/actions/run-contract-update — déclencher manuellement le job
router.post('/actions/run-contract-update', requirePermission('settings:write'), async (_req: AuthRequest, res: Response): Promise<void> => {
  try {
    const { runContractStatusUpdate } = await import('../scheduler')
    const result = await runContractStatusUpdate()
    res.json({ success: true, data: result })
  } catch (err) { handleRouteError(err, res) }
})

export default router
export { DEFAULTS as SETTING_DEFAULTS }